requirements.txt
selenium
bs4
python-dotenv
pandas
